backend/requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
django-filter>=24.3,<25
django-cors-headers>=4.4,<5
openpyxl>=3.1,<4
